Fixed issue with driver and added tests for non-zero page and indirect
Currently, there appears to be a bug in the indirect addressing with
zero page enabled. It is using the current page of the code when it
should be the zero page so a parameter was added to make it disable
that test with a warning until it can be investigated.